Refactor BugStorage with formatting and minor improvements
Refactored the `BugStorage` class to include a readonly modifier for `_converterSettings`. Reformatted the code to improve readability by adjusting brace placement in try-catch blocks and condensed single-line conditional statements. These changes enhance the maintainability and readability of the codebase.

+++ src/Gml.Core/Core/Helpers/BugTracker/BugStorage.cs
@@ -12,7 +12,7 @@ namespace Gml.Core.Helpers.BugTracker;
 
 public class FileStorageService(string filePath)
 {
-    private JsonSerializerSettings _converterSettings = new()
+    private readonly JsonSerializerSettings _converterSettings = new()
     {
         Converters =
         [
@@ -31,7 +31,9 @@ public class FileStorageService(string filePath)
         catch (Exception ex)
         {
             Console.WriteLine($"Error saving bug: {ex.Message}");
-        }finally{
+        }
+        finally
+        {
             Debug.WriteLine($"[{bugInfo.SendAt}] Error saving bug ");
         }
     }
@@ -52,10 +54,7 @@ public class FileStorageService(string filePath)
                     continue;
 
                 var bug = JsonConvert.DeserializeObject<BugInfo>(line, _converterSettings);
-                if (bug != null)
-                {
-                    bugs.Add(bug);
-                }
+                if (bug != null) bugs.Add(bug);
             }
         }
         catch (Exception ex)
@@ -106,10 +105,7 @@ public class FileStorageService(string filePath)
         }
         finally
         {
-            if (File.Exists(tempFilePath))
-            {
-                File.Delete(tempFilePath);
-            }
+            if (File.Exists(tempFilePath)) File.Delete(tempFilePath);
         }
     }
 }
